fix: Resolve broken AuthPage import
This change updates a broken import and consolidates the old AuthPage into a new [task].tsx route.
