+++ pages/app/auth/[task].tsx
@@ -1,7 +1,8 @@
+import { useRouter } from 'next/router';
 import React from 'react';
-import AuthCard from '../../../nebula-web/components/auth/AuthCard';
-import MarketingHeader from '../../../nebula-web/components/common/MarketingHeader';
-import { useAuthContext } from './auth-context';
+import AuthCard from '../../../components/auth/AuthCard';
+import MarketingHeader from '../../../components/common/MarketingHeader';
+import { useAuthContext } from '../../../modules/auth/auth-context';
 
 /**
  * A page that presents a sign-in/sign-up box to the user.
@@ -10,7 +11,22 @@ export default function AuthPage(): JSX.Element {
   // history.push('/app');
   // TODO: Like actually sign in
 
-  const { resetPassword, signInWithGoogle } = useAuthContext();
+  const { resetPassword, signInWithGoogle, signOut } = useAuthContext();
+
+  const router = useRouter();
+  const { task } = router.query;
+
+  const shouldSignOut = task === 'signOut';
+  if (shouldSignOut) {
+    signOut()
+      .then(() => {
+        console.debug('Succesfully signed out.');
+        router.push('/');
+      })
+      .catch((error) => {
+        console.error('Error signing out.', error);
+      });
+  }
 
   const handleForgetPassword = (email: string) => {
     resetPassword(email);
